+++ ts/packages/knowPro/src/search.ts
@@ -1,21 +1,21 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { mathLib, TopNCollection } from "typeagent";
+import { ITermToSemanticRefIndex, ScoredSemanticRef } from "./dataFormat.js";
 import {
-    ITermToSemanticRefIndex,
-    ScoredSemanticRef,
-    SemanticRefIndex,
-} from "./dataFormat.js";
+    QueryEvalContext,
+    SelectTopNSemanticRefsExpr,
+    TermsMatchExpr,
+} from "./query.js";
 
 export class SearchResult {
     constructor(
-        public matchedTerms: string[] = [],
-        public matchedSemanticRefs: ScoredSemanticRef[] = [],
+        public termMatches: string[] = [],
+        public semanticRefMatches: ScoredSemanticRef[] = [],
     ) {}
 
     public get hasMatches(): boolean {
-        return this.matchedTerms.length > 0;
+        return this.semanticRefMatches.length > 0;
     }
 }
 
@@ -24,147 +24,11 @@ export function searchTermsInIndex(
     terms: string[],
     maxMatches?: number,
 ): SearchResult {
-    const [matchedTerms, matchTable] = matchTermsInIndex(
-        semanticRefIndex,
-        terms,
+    const context = new QueryEvalContext();
+    const query = new SelectTopNSemanticRefsExpr(
+        new TermsMatchExpr(semanticRefIndex, terms),
+        maxMatches,
     );
-    if (matchTable.numMatches <= 0) {
-        return new SearchResult();
-    }
-
-    const minHitCount = matchedTerms.length;
-    return new SearchResult(
-        matchedTerms,
-        matchTable.getTopNScoring(maxMatches, minHitCount),
-    );
+    const [termMatches, semanticRefMatches] = query.eval(context);
+    return new SearchResult(termMatches, semanticRefMatches);
 }
-
-function matchTermsInIndex(
-    semanticRefIndex: ITermToSemanticRefIndex,
-    terms: string[],
-): [string[], MatchTable] {
-    const matchedTerms: string[] = [];
-    const matchTable = new MatchTable();
-    for (const term of terms) {
-        const scoredRefs = semanticRefIndex.lookupTerm(term);
-        if (scoredRefs && scoredRefs.length > 0) {
-            matchedTerms.push(term);
-            matchTable.add(scoredRefs);
-        }
-    }
-    return [matchedTerms, matchTable];
-}
-
-class MatchTable {
-    constructor(public matches = new Map<SemanticRefIndex, Match>()) {}
-
-    public get numMatches(): number {
-        return this.matches.size;
-    }
-
-    public get(refIndex: SemanticRefIndex): Match | undefined {
-        return this.matches.get(refIndex);
-    }
-
-    public add(matches: ScoredSemanticRef | ScoredSemanticRef[]) {
-        if (Array.isArray(matches)) {
-            for (const match of matches) {
-                this.addMatch(match);
-            }
-        } else {
-            this.addMatch(matches);
-        }
-    }
-
-    private addMatch(match: ScoredSemanticRef) {
-        let hit = this.matches.get(match.semanticRefIndex);
-        if (hit) {
-            hit.hitCount += 1;
-            hit.score += match.score;
-        } else {
-            hit = {
-                semanticRefIndex: match.semanticRefIndex,
-                hitCount: 1,
-                score: match.score,
-            };
-            this.matches.set(match.semanticRefIndex, hit);
-        }
-    }
-
-    public getSortedByScore(minHitCount = 0): ScoredSemanticRef[] {
-        if (this.matches.size === 0) {
-            return [];
-        }
-        const matches = [...this.matchesWithMinHitCount(minHitCount)];
-        matches.sort((x, y) => y.score - x.score);
-        return matches;
-    }
-
-    /**
-     * Return all matches with the 'top' or maximum score.
-     * @returns
-     */
-    public getTopScoring(): ScoredSemanticRef[] {
-        if (this.matches.size === 0) {
-            return [];
-        }
-        let maxScore = mathLib.max(
-            this.matches.values(),
-            (v) => v.score,
-        )!.score;
-        return [...this.filterMatches((match) => match.score === maxScore)];
-    }
-
-    /**
-     * Return the top N scoring matches
-     * @param maxMatches
-     * @returns
-     */
-    public getTopNScoring(
-        maxMatches?: number,
-        minHitCount = 0,
-    ): ScoredSemanticRef[] {
-        if (this.matches.size === 0) {
-            return [];
-        }
-        if (maxMatches) {
-            const topList = new TopNCollection(maxMatches, -1);
-            for (const match of this.matchesWithMinHitCount(minHitCount)) {
-                topList.push(match.semanticRefIndex, match.score);
-            }
-            const ranked = topList.byRank();
-            return ranked.map((m) => {
-                return {
-                    semanticRefIndex: m.item,
-                    score: m.score,
-                };
-            });
-        } else {
-            return this.getSortedByScore(minHitCount);
-        }
-    }
-
-    public *filterMatches(
-        predicate: (match: Match) => boolean,
-    ): IterableIterator<Match> {
-        for (const match of this.matches.values()) {
-            if (predicate(match)) {
-                yield match;
-            }
-        }
-    }
-
-    public matchesWithMinHitCount(
-        minHitCount?: number,
-    ): IterableIterator<Match> {
-        return minHitCount && minHitCount > 0
-            ? this.filterMatches((m) => m.hitCount >= minHitCount)
-            : this.matches.values();
-    }
-}
-
-type Match = {
-    semanticRefIndex: SemanticRefIndex;
-    hitCount: number;
-    score: number;
-};
